requirements.txt
streamlit
pandas
yfinance
pyyaml
reportlab
